ase agregan validaciones

+++ Servicio/servicio.py
@@ -7,27 +7,37 @@ from Servicio.Exception_api import *
 import requests
 from Servicio.Internal_errors import CodeInternalError
 
-
-#add notaMateria
+#region add materia
 def addNotaMateria(request): 
-    #notamateria=setNotaMateria(request)
-    #alumnoID=request.json['alumno_id']
-    #notamateriaID=request.json['nombremateria']
-    #notamateria_comp=NotaMateria.buscarNotaMateriaByNombreMateria(notamateriaID,alumnoID)
-    #if notamateria_comp.nombremateria is not none
-    #    if notamateria_comp.nombremateria==notamateriaID:
-    #        raise BadResquest('Ya existe la materia', CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)       
-    #    else:
+    notamateria=setNotaMateria(request)
+    if (existsNombreMateriaToAlumnoID(notamateria.alumno_fk,notamateria.nombremateria)):
+        raise Conflict('Existe una nota para esta materia', CodeInternalError.ERROR_INTERNAL_14_REQUEST_DATA_DUPLICATED)
+    else:
+        try:
+            notamateria.save()
+            content = {'detalle': 'Registro creado.'}
+            return content, status.HTTP_201_CREATED 
+        except Exception as identifier:
+            raise InternalServerError('Error relacionado con base de datos', CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)              
+#endregion 
+
+#region get all notas manteria by ALUMNO ID
+def getNotasMateriasByAlumnoID(alumnoid):
+    if not(isNumber(alumnoid)):
+        raise BadResquest('AlumnoID es un tipo invalido', CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
     try:
-        notamateria.save()
-        content = {'detalle': 'Recurso creado.'}
-        return content, status.HTTP_201_CREATED 
+        obj=NotaMateria.getNotasMateriasByAlumnoID(alumnoid)
     except Exception as identifier:
-        raise InternalServerError('Error relacionado con base de datos.', CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)              
-    
-#get
-def findNotasMateriasByAlumnoID(id):
-    print('antes del try')
+        raise InternalServerError(identifier, CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)          
+    if obj is not None and len(obj)!=0:
+        json_Str=jsonify([e.serializar() for e in obj]) 
+        return json_Str
+    else:
+        raise BadResquest('Recurso no encontrado en la base de datos.',CodeInternalError.ERROR_INTERNAL_12_REQUEST_NOT_FOUND)
+#endregion
+
+#region get one notamanteriaID to an AlumnoID
+def getNotasMateriasToAlumnoIDbyNotaMateriaID(id):
     try:
         obj=NotaMateria.getNotasMateriasByAlumnoID(id)
         print('hola')
@@ -38,22 +48,25 @@ def findNotasMateriasByAlumnoID(id):
         return json_Str
     else:
         raise BadResquest('Recurso no encontrado en la base de datos.',CodeInternalError.ERROR_INTERNAL_12_REQUEST_NOT_FOUND)
-            
+#endregion
 
-# Get all NotasMaterias
+
+
+
+#region Get all NotasMaterias
 def getNotasMaterias():
     materias=NotaMateria.buscarNotasMaterias()
     json_Str=jsonify([e.serializar() for e in materias]) 
     return json_Str
-    #page=NotaMateria.buscarNotasMaterias()
-    #final_list = []
-    #for element in page:
-    #    final_list.append(element)
-    #return jsonify(final_list), status_code
-
+    '''page=NotaMateria.buscarNotasMaterias()
+    final_list = []
+    for element in page:
+       final_list.append(element)
+    return jsonify(final_list), status_code'''
+#endregion
     
 
-#Update
+#region Update notamateria
 def updateNotaMateria(request): 
     try:
         notamateriaID=request.json['notamateria_id']
@@ -63,29 +76,26 @@ def updateNotaMateria(request):
     except Exception as identifier:
         raise BadResquest('Estructa de Json incorrecta',CodeInternalError.ERROR_INTERNAL_10_JSON_BAD_FORMED) 
     #se valida el nro de la nota
-    if (int(notafinal)>=11 or int(notafinal)<0):
-            raise BadResquest('Nota invalida',CodeInternalError.ERROR_INTERNAL_10_JSON_BAD_FORMED)
+    isValidNotaFinal(notafinal)
+    #if (int(notafinal)>=11 or int(notafinal)<0):
+    #        raise BadResquest('Nota invalida',CodeInternalError.ERROR_INTERNAL_10_JSON_BAD_FORMED)
         #Se valida si la notamateriaID recibida corresponde al alumno en cuestión. Valida si el user no existe.
-    notamateria=NotaMateria.buscarNotaMateriaByNotamateriaID(notamateriaID)
+    notamateria=NotaMateria.getNotaMateriaByNotamateriaID(notamateriaID)
     if notamateria is None:
         raise BadResquest('Los datos recibidos no coinciden.',CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
     if (int(alumnoID)==int(notamateria.alumno_fk)):
         try:
             notamateria.nombremateria=nombremateria
             notamateria.notafinal=notafinal
-            '''print(notamateria.nombremateria)
-            print(notamateria.notafinal)
-            print(notamateria.alumno_fk)
-            print(notamateria.notamateria_id)'''
             notamateria.save()
             return ('Recurso actualizado.')
         except Exception as identifier:
             raise InternalServerError('Error relacionado con base de datos.', CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)       
     else:
         raise BadResquest('Los datos recibidos no coinciden.',CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED) 
-        
+#endregion    
     
-# Update
+#region Update notamateria byAlumnoID
 def updateNotaMateriasByAlumnoID(id):
     try:      
         obj=NotaMateria.buscarMateriasByAlumnoID(id)
@@ -93,47 +103,85 @@ def updateNotaMateriasByAlumnoID(id):
         return (json_str)
     except Exception as e:
 	    return(str(e))
- 
-#
-def deleteNotaMateria(id):
+#endregion 
+
+#region delete notamaterias
+def deleteNotaMateria(alumnoid,notamateriaid):
+    if not(isNumber(alumnoid)):
+        raise BadResquest('AlumnoID es un tipo invalido', CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
+    elif not(isNumber(notamateriaid)):
+        BadResquest('NotamateriaID es un tipo invalido', CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
+    if (not(existsNotaMateriaIDToAlumnoID(alumnoid,notamateriaid))):
+        raise BadResquest('No existe la materia para el alumno asociado', CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)
     try:
-        notamateria=NotaMateria.buscarNotaMateriaByNotamateriaID(id)
+        notamateria=NotaMateria.getNotaMateriaByNotamateriaID(notamateriaid)
     except Exception as identifier:
         raise InternalServerError('Error relacionado con base de datos.', CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)          
     if notamateria is not None:
         try:
             notamateria.delete()
-            return ('Recurso eliminado.',status.HTTP_200_OK)
+            return ('',status.HTTP_204_NO_CONTENT)
         except Exception as identifier:
             raise InternalServerError('Error relacionado con base de datos.', CodeInternalError.ERROR_INTERNAL_11_CONEXION_BD)   
     else:
         raise BadResquest('Recurso no encontrado en la base de datos.',CodeInternalError.ERROR_INTERNAL_12_REQUEST_NOT_FOUND)  
-     
+#endregion     
 
-#Methods
+#region Common Methods
 def setNotaMateria(request):
-    try:   
-        notamateria=NotaMateria(
-            request.json['alumno_id'],
-            request.json['nombremateria'],
-            request.json['notafinal'] 
-        )
+    try:  
+        request.get_json()
     except Exception as identifier:
-        raise BadResquest('Estructa de Json incorrecta',CodeInternalError.ERROR_INTERNAL_10_JSON_BAD_FORMED)
+        raise BadResquest('Estructura de archivo invalida',CodeInternalError.ERROR_INTERNAL_10_JSON_BAD_FORMED)
+    isValidDataType(request)
+    isValidNotaFinal(request.json['notafinal']) 
+    notamateria=NotaMateria(
+        request.json['alumno_id'],
+        request.json['nombremateria'],
+        request.json['notafinal']
+        )
     return notamateria
-       
 
-       
-#def setMessajeFormatJson():
-#    return ({'detail':'Estructura json no soportada'},status.HTTP_400_BAD_REQUEST)
-#
+def isValidDataType(request):
+    if (str((request.json['nombremateria'])).isnumeric()):
+        raise BadResquest('Nombre de materia no es un tipo de dato valido',CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
+    elif (not(isNumber((request.json['notafinal'])))):
+        raise BadResquest('Nota final no es un tipo de dato valido',CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
+    elif (not(isNumber((request.json['alumno_id'])))):
+        raise BadResquest('Alumno id no es un tipo de dato valido',CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
 
-def imprimirJson():
-    notamateria = NotaMateria('11','estadistica','10') 
-    json_data = json.dumps(notamateria) 
-    print (json_data)
-    return (json_data)
 
+''' Valida que sea entero'''         
+def isNumber(json_value):
+    try:
+        int(json_value)
+        return True
+    except:
+        return False
+
+def existsNotaMateriaIDToAlumnoID(alumnoid, notamateriaid):
+    notamateria_comp=NotaMateria.getNotaMateriaByNombreMateria(notamateriaid,alumnoid)
+    dir(notamateria_comp)
+    if notamateria_comp is None:
+        return False
+    else:
+        return True
+
+def existsNombreMateriaToAlumnoID(alumnoid, nombremateria):
+    notamateria_comp=NotaMateria.getNotaMateriaByNombreMateria(alumnoid,nombremateria)
+    dir(notamateria_comp)
+    if notamateria_comp is None:
+        return False
+    else:
+        return True
+#endregion    
+
+
+#region bussiness rules
+def isValidNotaFinal(notafinal):
+    if (int(notafinal)>=11 or int(notafinal)<0):
+        raise BadResquest('Nota invalida',CodeInternalError.ERROR_INTERNAL_13_REQUEST_DATA_NOT_MATCHED)
+#endregion
